Separate Config logic from invoke logic.
