+++ MarvelApiClient/src/main/java/com/karumi/MarvelApiClient/MarvelApiClient.java
@@ -16,42 +16,21 @@
 package com.karumi.marvelapiclient;
 
 import com.google.gson.Gson;
-import com.squareup.okhttp.OkHttpClient;
-import com.squareup.okhttp.logging.HttpLoggingInterceptor;
 import java.io.IOException;
 import retrofit.Call;
-import retrofit.GsonConverterFactory;
 import retrofit.Response;
-import retrofit.Retrofit;
 
-/**
- * Api client for access to Marvel Api.
- */
-public class MarvelApiClient {
-
+class MarvelApiClient {
   private static final int INVALID_AUTH_CODE = 401;
-  private static MarvelApiClient singleton;
-  private final String publicKey;
-  private final String privateKey;
-  private final boolean debug;
-  private final ApiCall apiCall;
 
-  MarvelApiClient(String publicKey, String privateKey, ApiCall apiCall, boolean debug) {
-    this.publicKey = publicKey;
-    this.privateKey = privateKey;
-    this.apiCall = apiCall;
-    this.debug = debug;
-  }
+  private final MarvelApiConfig marvelApiConfig;
 
-  public static MarvelApiClient with(String publicKey, String privateKey) {
-    if (singleton == null) {
-      singleton = new Builder(publicKey, privateKey).build();
-    }
-    return singleton;
+  public MarvelApiClient(MarvelApiConfig marvelApiConfig) {
+    this.marvelApiConfig = marvelApiConfig;
   }
 
   <T> T getApi(Class<T> apiRest) {
-    return apiCall.obtainApi(apiRest);
+    return marvelApiConfig.getRetrofit().create(apiRest);
   }
 
   public <T> T execute(Call<T> call) throws MarvelApiException {
@@ -90,83 +69,6 @@ public class MarvelApiClient {
       throw new MarvelAuthApiException(execute.code(), marvelCode, marvelDescription, null);
     } else {
       throw new MarvelApiException(execute.code(), marvelCode, marvelDescription, null);
-    }
-  }
-
-  /**
-   * Fluent API for creating {@link MarvelApiClient} instances.
-   */
-  @SuppressWarnings("UnusedDeclaration") public static class Builder {
-
-    private static final String MARVEL_API_BASE_URL = "http://gateway.marvel.com/v1/public/";
-    private final String privateKey;
-    private final String publicKey;
-    private boolean debug;
-    private Retrofit retrofit;
-    private String baseUrl = MARVEL_API_BASE_URL;
-    private TimeProvider timeProvider = new TimeProvider();
-
-    public Builder(String publicKey, String privateKey) {
-      if (publicKey == null) {
-        throw new IllegalArgumentException("publicKey must not be null.");
-      }
-
-      if (privateKey == null) {
-        throw new IllegalArgumentException("privateKey must not be null.");
-      }
-
-      this.publicKey = publicKey;
-      this.privateKey = privateKey;
-    }
-
-    public Builder debug() {
-      this.debug = true;
-      return this;
-    }
-
-    public Builder baseUrl(String url) {
-      this.baseUrl = url;
-      return this;
-    }
-
-    public Builder retrofit(Retrofit retrofit) {
-      if (retrofit == null) {
-        throw new IllegalArgumentException("retrofit must not be null.");
-      }
-      this.retrofit = retrofit;
-      return this;
-    }
-
-    public MarvelApiClient build() {
-      if (retrofit == null) {
-        retrofit = createDefaultRetrofit(baseUrl, debug);
-      }
-
-      addKeysToParams(retrofit, publicKey, privateKey, timeProvider);
-
-      return new MarvelApiClient(publicKey, privateKey, new ApiCall(retrofit), debug);
-    }
-
-    private void addKeysToParams(Retrofit retrofit, String publicKey, String privateKey,
-        TimeProvider timeProvider) {
-      OkHttpClient client = retrofit.client();
-      client.interceptors().add(new AuthInterceptor(publicKey, privateKey, timeProvider));
-    }
-
-    private Retrofit createDefaultRetrofit(String baseUrl, boolean debug) {
-      OkHttpClient client = new OkHttpClient();
-      if (debug) {
-        HttpLoggingInterceptor interceptor = new HttpLoggingInterceptor();
-        interceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
-        client.interceptors().add(interceptor);
-      }
-
-      Retrofit retrofit = new Retrofit.Builder().baseUrl(baseUrl)
-          .client(client)
-          .addConverterFactory(GsonConverterFactory.create())
-          .build();
-
-      return retrofit;
     }
   }
 }
