+++ MarvelApiClient/src/test/java/com/karumi/MarvelApiClient/MarvelApiConfigTest.java
@@ -17,23 +17,23 @@ package com.karumi.marvelapiclient;
 
 import org.junit.Test;
 
-public class MarvelApiClientTest {
+public class MarvelApiConfigTest {
 
   private static final String ANY_PRIVATE_KEY = "PRIVATEKEY";
   private static final String ANY_PUBLIC_KEY = "PUBLICKEY";
 
   @Test(expected = IllegalArgumentException.class)
   public void shouldThrowInvalidArgumentExceptionWhenCreateWithoutPublicKey() {
-    MarvelApiClient.with(null, ANY_PRIVATE_KEY);
+    MarvelApiConfig.with(null, ANY_PRIVATE_KEY);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void shouldThrowInvalidArgumentExceptionWhenCreateWithoutPrivateKey() {
-    MarvelApiClient.with(ANY_PUBLIC_KEY, null);
+    MarvelApiConfig.with(ANY_PUBLIC_KEY, null);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void shouldThrowInvalidArgumentExceptionWhenCreateWithInvalidRetrofit() {
-    new MarvelApiClient.Builder(ANY_PUBLIC_KEY, ANY_PRIVATE_KEY).retrofit(null);
+    new MarvelApiConfig.Builder(ANY_PUBLIC_KEY, ANY_PRIVATE_KEY).retrofit(null);
   }
 }
